+++ CHANGELOG.md
@@ -6,9 +6,11 @@
   - Removed .toObject()
 - Removed embedded API GQL concept
 - Removed Directives [embedApi, resolve, @value]
-- Removed toId Transform
-- Removed Model.tform()
+- Removed toId Transform -> use @field(id: '')
+- Removed Model.tform() -> use Model.hydrate(data, context)
+- Removed
 - Changed Rule + Transform API
+- Added .resolve() terminal command
 
 ## v0.9.x
 - Subscriptions API